More options for the MC simulation script

+++ MC/do_simulate.sh
@@ -3,10 +3,12 @@
 ##############################################################################
 #
 # Create a sub-directory YYYYMMDD-hhmm"
-# Execute a full SNEMO/Falaise simulation-->reconstruction-->sentitivity
+# Execute a full SNEMO/Falaise simulation-->reconstruction-->sensitivity
 #
 # Usage: ./do_simulate.sh <configuration>
 # Output: configuration.root
+#
+# Can use binaries from a built version of Falaise or a Docker image.
 #
 # From Falaise: the resultant file <configuration>-reco.root contains
 # a single flat TTree structure that may be browsed interactively.
@@ -15,9 +17,13 @@
 #
 ##############################################################################
 
-SENSITIVITY="$HOME/SuperNEMO/SensitivityModuleBuild"
+PURGE=false                                                                     # delete sim and reco files to save space
+SENSITIVITY="$HOME/SuperNEMO/SensitivityModuleBuild"                            # location of the SensitivityModule
+RECO_CONFIG="@falaise:snemo/demonstrator/reconstruction/official-2.0.0.conf"    # reconstruction config file
+FALAISE="/vols/build/snemo/Falaise.build/bin/"                                  # Falaise software
+#FALAISE="singularity exec --home /home/hep/pfranchi --bind /vols/snemo/ -c /vols/build/snemo/falaise/falaise_latest.sif "
 
-###############################
+##############################################################################
 
 shopt -s expand_aliases
 alias snemo-sing='singularity exec --home /home/hep/pfranchi --bind /vols/snemo/ -c /vols/build/snemo/falaise/falaise_latest.sif'
@@ -33,7 +39,7 @@ WHERE=$PWD/`date "+%Y%m%d-%H%M"`
 mkdir -p $WHERE
 rm -f $WHERE/$CONFIGURATION.log
 
-snemo-sing flsimulate --version >> $WHERE/$CONFIGURATION.log
+${FALAISE}flsimulate --version >> $WHERE/$CONFIGURATION.log
 
 # Create Sensitivity conf with the proper filename
 cp ${SENSITIVITY}/SensitivityModuleExample.conf $WHERE/$CONFIGURATION-sensitivity.conf
@@ -43,18 +49,21 @@ echo "Using: " ${CONFIGURATION}.conf
 echo "Working in: " $WHERE
 
 echo "Simulation..."
-snemo-sing flsimulate -c $PWD/$CONFIGURATION.conf -o $WHERE/$CONFIGURATION-sim.brio &>> $WHERE/$CONFIGURATION.log
-echo "snemo-sing flsimulate -c $PWD/$CONFIGURATION.conf -o $WHERE/$CONFIGURATION-sim.brio"
+${FALAISE}flsimulate -c $PWD/$CONFIGURATION.conf -o $WHERE/$CONFIGURATION-sim.brio &>> $WHERE/$CONFIGURATION.log
+#echo "${FALAISE}flsimulate -c $PWD/$CONFIGURATION.conf -o $WHERE/$CONFIGURATION-sim.brio"
 
 if [ ! -f "$WHERE/$CONFIGURATION-sim.brio" ]; then
-    echo "Simulation not present, something wrong with flsimulate!!!"
+    echo -e "Simulation not present, something wrong with flsimulate!!!\n"
+    tail -n +12 $WHERE/$CONFIGURATION.log
     exit
 fi
 
 echo "Reconstruction..."
-snemo-sing flreconstruct -i $WHERE/$CONFIGURATION-sim.brio -p @falaise:snemo/demonstrator/reconstruction/official-2.0.0.conf -o $WHERE/$CONFIGURATION-reco.brio &>> $WHERE/$CONFIGURATION.log
+${FALAISE}flreconstruct -i $WHERE/$CONFIGURATION-sim.brio -p ${RECO_CONFIG} -o $WHERE/$CONFIGURATION-reco.brio &>> $WHERE/$CONFIGURATION.log
 # Remove unused filed to save space
-rm -f $WHERE/$CONFIGURATION-sim.brio
+if [ "$PURGE" = true ] ; then
+    rm -f $WHERE/$CONFIGURATION-sim.brio
+fi
 
 if [ ! -f "$WHERE/$CONFIGURATION-reco.brio" ]; then
     echo "Reconstruction not present, something wrong with flreconstruct!!!"
@@ -70,7 +79,9 @@ if [ ! -f "$WHERE/$CONFIGURATION.root" ]; then
 fi
 
 # Remove unused filed to save space   
-rm -f $WHERE/$CONFIGURATION-reco.brio
-rm -f $WHERE/$CONFIGURATION-reco.root
+if [ "$PURGE" = true ] ; then
+    rm -f $WHERE/$CONFIGURATION-reco.brio
+    rm -f $WHERE/$CONFIGURATION-reco.root
+fi
 
 echo "--> Output from Sensitivity in: "$WHERE/$CONFIGURATION.root
